+++ CHANGES.md
@@ -9,6 +9,7 @@ Changelog
 - Add plone.formwidget.geolocation >= 3.0.5 dependency. [sverbois]
 - Add a Javascript workaround to fix "Add item to default page". [sverbois]
 - Use Python 3.11 and Plone 6.0.6 in development buildout. [sverbois]
+- Add Newspage content type. [sverbois]
 
 1.0.0a15 (2023-01-01)
 ---------------------